Module Api.Commit

A specific Git commit.

type t

val id : t -> Current_git.Commit_id.t

Identifier for a specific Git commit, which can be used to fetch it.

val set_status : t Current.t -> string -> Status.t Current.t -> unit Current.t

set_status commit context status sets the status of commit/context to status.

val owner_name : t -> string

owner_name t is the "owner/name" string identifying the repository.

val repo_id : t -> Repo_id.t

Construct a Repo_id.t that t belongs to.

val hash : t -> string

hash t is the Git commit hash of t.

val committed_date : t -> string

committed_date t is the datetime when t was committed.

val message : t -> string

message t is the Git commit message of t.

val pp : t Fmt.t

Pretty print t.

val compare : t -> t -> int

val uri : t -> Uri.t

uri t is a URI for the GitLab web page showing t.

val mr_name : t -> string option

mr_name t is the name of the ref that the commit belongs to if it is a MR, and None if it is a branch.

val branch_name : t -> string option

branch_name t is the name of the ref that the commit belongs to if it is a branch, and None if it is a PR.