Module Api.Ref

A Ref as an indirect way of referring to a commit.

type mr_info = {

id : int;

base : string;

title : string;

body : string;

}

type t = [

| `Ref of string

| `MR of mr_info

]

t is a regular git ref or a merge request record of mr_info.

type id = [

| `Ref of string

| `MR of int

]

Id as a regular git ref or a merge request.

This is intended to be used by a user when a t type is manipulated by the API.

val pp : t Fmt.t

Pretty print t.

val compare : t -> t -> int

val to_git : t -> string

to_git t is the Git-format string of the ref, e.g."refs/pull/%d/head"